+++ crates/cli/src/app.rs
@@ -54,6 +54,15 @@ pub enum Commands {
     },
 
     #[command(
+        name = "clean",
+        about = "Clean the ~/.proto directory by removing stale and old tools."
+    )]
+    Clean {
+        #[arg(long, help = "Clean tools older than the specified number of days")]
+        days: Option<u8>,
+    },
+
+    #[command(
         name = "completions",
         about = "Generate command completions for your current shell."
     )]